+++ Domain/DTOs/Requests/UpdateRequestDtos/UpdateResultRequestDto.cs
@@ -5,7 +5,7 @@ namespace App.Core.DTOs.Requests.UpdateRequestDtos
     public class UpdateResultRequestDto
     {
         public Guid ExaminationId { get; set; }
-        public int? TotalScore { get; set; }
+        public double? TotalScore { get; set; }
         [Column(TypeName = "json")]
         public string? Breakdown { get; set; }
     }
